ml-serviceace/requirements.txt
fastapi
uvicorn
requests
pydantic
python-dotenv
openai
